use f.Truncate to change size of the file

+++ randimgmain/randimgmain.go
@@ -15,7 +15,7 @@ var (
 	filename string
 	seq      int
 	picfmt   string
-	fixMib   int
+	fixMib   int64
 	many     int
 )
 
@@ -23,7 +23,7 @@ func init() {
 	flag.IntVar(&width, "w", 640, "picture width")
 	flag.IntVar(&height, "h", 320, "picture height")
 	flag.IntVar(&seq, "s", 0, "picture sequence number")
-	flag.IntVar(&fixMib, "m", 0, "fixed size(MiB)")
+	flag.Int64Var(&fixMib, "m", 0, "fixed size(MiB)")
 	flag.IntVar(&many, "i", 1, "how many pictures to create")
 	flag.StringVar(&picfmt, "f", "png", "picture format(png/jpg)")
